+++ projects/angular-cesium/src/lib/angular-cesium-widgets/models/polygon-edit-options.ts
@@ -5,7 +5,6 @@ export interface PolygonProps {
   fill?: boolean;
   classificationType?: any;
   zIndex?: any;
-  useGroundPrimitiveOutline?: boolean;
 }
 
 export interface PolygonEditOptions extends PolylineEditOptions {